improved use of ASSET_LOC to be handled by CMake
I can forsee this being an issue when running `make check` in release
mode. This is because the ASSET_LOC will be set to the install
location of the data directory, and not the local build version. When
any test goes to use the installed data directory it will fail because
the files won't be there. This needs to be fixed before v0.3.0 is released.