Moved to C++11 again however we are now picking standard per build
What this means is that when we decide to use a C++14 feature we only
will build the specified module at that standard.

Please note: If you do decide to use a C++14 feature, becuase it is
not the current standard we should account for older compilers that
may not have C++14, and must still be able to build accordingly.