updated data install location to better conform to OS specifications
Specifically, on Linux all Architecture-independent (shared) data (i.e
our Data folder) should end up here. The standard convention for this
is /usr/share/{package}/ so that is what we will conform to.

Windows does not spread an install across a filesystem, like Linux
does, so we can just plop it in the install location.

This also resolves an install conflict with VSCode, and other possible
programs, where they would check for write access to /usr/data (a
non-standard folder that we were attempting to use). This lead to an
issue where, if we are install on the system, the /usr/data folder is
created, and ownership is given to root. Unless VSCode is then run as
root (NOT suggested), it becomes unable to launch.

VSCode is my primary editor beyond EMACS. OOPS! 😋