renamed Utils library to Core and added core/ subfolder in src/
this also moved utils.cpp and xml.cpp into the core folder and added a
CMakeLists.txt file to src/core/

with the addition of the xml loading to the Core namespace, it felt
appropriate to rename the Utils library to Core, as Core is now the
larger namespace. I do not know if XMl belongs in the Core namespace,
but that is where it will be for now. I also updated comments
throughout the project to reflect this change.